+++ imports/client/components/HuntListPage.tsx
@@ -2,9 +2,7 @@
 import { Meteor } from 'meteor/meteor';
 import { useSubscribe, useTracker } from 'meteor/react-meteor-data';
 import { faEdit } from '@fortawesome/free-solid-svg-icons/faEdit';
-import { faInfo } from '@fortawesome/free-solid-svg-icons/faInfo';
 import { faMinus } from '@fortawesome/free-solid-svg-icons/faMinus';
-import { faPlus } from '@fortawesome/free-solid-svg-icons/faPlus';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, {
   MouseEvent, useCallback, useImperativeHandle, useRef, useState,
@@ -12,513 +10,13 @@ import React, {
 import Alert from 'react-bootstrap/Alert';
 import Button from 'react-bootstrap/Button';
 import ButtonGroup from 'react-bootstrap/ButtonGroup';
-import Col from 'react-bootstrap/Col';
-import FormCheck from 'react-bootstrap/FormCheck';
-import FormControl, { FormControlProps } from 'react-bootstrap/FormControl';
-import FormGroup from 'react-bootstrap/FormGroup';
-import FormLabel from 'react-bootstrap/FormLabel';
-import FormText from 'react-bootstrap/FormText';
 import Modal from 'react-bootstrap/Modal';
-import Row from 'react-bootstrap/Row';
 import { Link } from 'react-router-dom';
 import Ansible from '../../Ansible';
-import DiscordCache from '../../lib/models/DiscordCache';
 import Hunts from '../../lib/models/Hunts';
-import Settings from '../../lib/models/Settings';
 import { userMayCreateHunt, userMayUpdateHunt } from '../../lib/permission_stubs';
-import { HuntType, SavedDiscordObjectType } from '../../lib/schemas/Hunt';
-import { SettingType } from '../../lib/schemas/Setting';
-import { useBreadcrumb } from '../hooks/breadcrumb';
+import { HuntType } from '../../lib/schemas/Hunt';
 import ModalForm, { ModalFormHandle } from './ModalForm';
-
-const splitLists = function (lists: string): string[] {
-  const strippedLists = lists.trim();
-  if (strippedLists === '') {
-    return [];
-  }
-
-  return strippedLists.split(/[, ]+/);
-};
-
-interface DiscordSelectorParams {
-  disable: boolean;
-  id: string;
-  value: SavedDiscordObjectType | undefined;
-  onChange: (next: SavedDiscordObjectType | undefined) => void;
-}
-
-interface DiscordSelectorProps extends DiscordSelectorParams {
-  loading: boolean;
-  options: SavedDiscordObjectType[];
-}
-
-const DiscordSelector = ({
-  disable, id: formId, value, onChange, loading, options,
-}: DiscordSelectorProps) => {
-  const onValueChanged: FormControlProps['onChange'] = useCallback((e) => {
-    if (e.currentTarget.value === 'empty') {
-      onChange(undefined);
-    } else {
-      const match = options.find((obj) => { return obj.id === e.currentTarget.value; });
-      if (match) {
-        onChange(match);
-      }
-    }
-  }, [onChange, options]);
-
-  const formOptions = useCallback((): SavedDiscordObjectType[] => {
-    // List of the options.  Be sure to include the saved option if it's (for
-    // some reason) not present in the channel list.
-    const noneOption = {
-      id: 'empty',
-      name: 'disabled',
-    } as SavedDiscordObjectType;
-
-    if (value) {
-      if (!options.find((opt) => {
-        return opt.id === value!.id;
-      })) {
-        return [noneOption, value, ...options];
-      }
-    }
-    return [noneOption, ...options];
-  }, [value, options]);
-
-  if (loading) {
-    return <div>Loading discord resources...</div>;
-  } else {
-    return (
-      <FormControl
-        id={formId}
-        as="select"
-        type="text"
-        placeholder=""
-        value={value && value.id}
-        disabled={disable}
-        onChange={onValueChanged}
-      >
-        {formOptions().map(({ id, name }) => {
-          return (
-            <option key={id} value={id}>{name}</option>
-          );
-        })}
-      </FormControl>
-    );
-  }
-};
-
-const DiscordChannelSelector = ({ guildId, ...rest }: DiscordSelectorParams & { guildId: string }) => {
-  const cacheLoading = useSubscribe('discord.cache', { type: 'channel' });
-  const loading = cacheLoading();
-
-  const { options } = useTracker(() => {
-    const discordChannels: SavedDiscordObjectType[] = DiscordCache.find({
-      type: 'channel',
-      'object.guild': guildId,
-      // We want only text channels, since those are the only ones we can bridge chat messages to.
-      'object.type': 'text',
-    }, {
-      // We want to sort them in the same order they're provided in the Discord UI.
-      sort: { 'object.rawPosition': 1 },
-      fields: { 'object.id': 1, 'object.name': 1 },
-    })
-      .map((c) => c.object as SavedDiscordObjectType);
-
-    return {
-      options: discordChannels,
-    };
-  }, [guildId]);
-  return (
-    <DiscordSelector
-      loading={loading}
-      options={options}
-      {...rest}
-    />
-  );
-};
-
-const DiscordRoleSelector = ({ guildId, ...rest }: DiscordSelectorParams & { guildId: string }) => {
-  const cacheLoading = useSubscribe('discord.cache', { type: 'role' });
-  const loading = cacheLoading();
-  const { options } = useTracker(() => {
-    const discordRoles: SavedDiscordObjectType[] = DiscordCache.find({
-      type: 'role',
-      'object.guild': guildId,
-      // The role whose id is the same as the guild is the @everyone role, don't want that
-      'object.id': { $ne: guildId },
-      // Managed roles are owned by an integration
-      'object.managed': false,
-    }, {
-      // We want to sort them in the same order they're provided in the Discord UI.
-      sort: { 'object.rawPosition': 1 },
-      fields: { 'object.id': 1, 'object.name': 1 },
-    })
-      .map((c) => c.object as SavedDiscordObjectType);
-
-    return {
-      options: discordRoles,
-    };
-  }, [guildId]);
-  return (
-    <DiscordSelector
-      loading={loading}
-      options={options}
-      {...rest}
-    />
-  );
-};
-
-export interface HuntModalSubmit {
-  // eslint-disable-next-line no-restricted-globals
-  name: string;
-  mailingLists: string[];
-  signupMessage: string;
-  openSignups: boolean;
-  hasGuessQueue: boolean;
-  homepageUrl: string;
-  submitTemplate: string;
-  puzzleHooksDiscordChannel: SavedDiscordObjectType | undefined;
-  firehoseDiscordChannel: SavedDiscordObjectType | undefined;
-  memberDiscordRole: SavedDiscordObjectType | undefined;
-}
-
-enum HuntModalFormSubmitState {
-  IDLE = 'idle',
-  SUBMITTING = 'submitting',
-  FAILED = 'failed',
-}
-
-type HuntModalFormHandle = {
-  show: () => void;
-}
-
-const HuntModalForm = React.forwardRef(({
-  hunt, onSubmit,
-}: {
-  hunt?: HuntType;
-  onSubmit: (state: HuntModalSubmit, callback: (error?: Error) => void) => void;
-}, forwardedRef: React.Ref<HuntModalFormHandle>) => {
-  useSubscribe('mongo.settings', { name: 'discord.guild' });
-  const guildId = useTracker(() => {
-    const setting = Settings.findOne({ name: 'discord.guild' }) as SettingType & { name: 'discord.guild' } | undefined;
-    return setting?.value.guild.id;
-  }, []);
-
-  const [submitState, setSubmitState] = useState<HuntModalFormSubmitState>(HuntModalFormSubmitState.IDLE);
-  const [errorMessage, setErrorMessage] = useState<string>('');
-
-  const [name, setName] = useState<string>(hunt?.name ?? '');
-  const [mailingLists, setMailingLists] = useState<string>(hunt?.mailingLists.join(', ') ?? '');
-  const [signupMessage, setSignupMessage] = useState<string>(hunt?.signupMessage ?? '');
-  const [openSignups, setOpenSignups] = useState<boolean>(hunt?.openSignups ?? false);
-  const [hasGuessQueue, setHasGuessQueue] = useState<boolean>(hunt?.hasGuessQueue ?? true);
-  const [homepageUrl, setHomepageUrl] = useState<string>(hunt?.homepageUrl ?? '');
-  const [submitTemplate, setSubmitTemplate] = useState<string>(hunt?.submitTemplate ?? '');
-  const [puzzleHooksDiscordChannel, setPuzzleHooksDiscordChannel] = useState<SavedDiscordObjectType | undefined>(hunt?.puzzleHooksDiscordChannel);
-  const [firehoseDiscordChannel, setFirehoseDiscordChannel] = useState<SavedDiscordObjectType | undefined>(hunt?.firehoseDiscordChannel);
-  const [memberDiscordRole, setMemberDiscordRole] = useState<SavedDiscordObjectType | undefined>(hunt?.memberDiscordRole);
-
-  const formRef = useRef<ModalFormHandle>(null);
-
-  useImperativeHandle(forwardedRef, () => ({
-    show: () => {
-      if (formRef.current) {
-        formRef.current.show();
-      }
-    },
-  }));
-
-  const onNameChanged: FormControlProps['onChange'] = useCallback((e) => {
-    setName(e.currentTarget.value);
-  }, []);
-
-  const onMailingListsChanged: FormControlProps['onChange'] = useCallback((e) => {
-    setMailingLists(e.currentTarget.value);
-  }, []);
-
-  const onSignupMessageChanged: FormControlProps['onChange'] = useCallback((e) => {
-    setSignupMessage(e.currentTarget.value);
-  }, []);
-
-  const onOpenSignupsChanged = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-    setOpenSignups(e.currentTarget.checked);
-  }, []);
-
-  const onHasGuessQueueChanged = useCallback((e: React.ChangeEvent<HTMLInputElement>) => {
-    setHasGuessQueue(e.currentTarget.checked);
-  }, []);
-
-  const onHomepageUrlChanged: FormControlProps['onChange'] = useCallback((e) => {
-    setHomepageUrl(e.currentTarget.value);
-  }, []);
-
-  const onSubmitTemplateChanged: FormControlProps['onChange'] = useCallback((e) => {
-    setSubmitTemplate(e.currentTarget.value);
-  }, []);
-
-  const onPuzzleHooksDiscordChannelChanged = useCallback((next: SavedDiscordObjectType | undefined) => {
-    setPuzzleHooksDiscordChannel(next);
-  }, []);
-
-  const onFirehoseDiscordChannelChanged = useCallback((next: SavedDiscordObjectType | undefined) => {
-    setFirehoseDiscordChannel(next);
-  }, []);
-
-  const onMemberDiscordRoleChanged = useCallback((next: SavedDiscordObjectType | undefined) => {
-    setMemberDiscordRole(next);
-  }, []);
-
-  const onFormSubmit = (callback: () => void) => {
-    setSubmitState(HuntModalFormSubmitState.SUBMITTING);
-    const sendState: HuntModalSubmit = {
-      name,
-      mailingLists: splitLists(mailingLists),
-      signupMessage,
-      openSignups,
-      hasGuessQueue,
-      homepageUrl,
-      submitTemplate,
-      puzzleHooksDiscordChannel,
-      firehoseDiscordChannel,
-      memberDiscordRole,
-    };
-
-    onSubmit(sendState, (error?: Error) => {
-      if (error) {
-        setErrorMessage(error.message);
-        setSubmitState(HuntModalFormSubmitState.FAILED);
-      } else {
-        setSubmitState(HuntModalFormSubmitState.IDLE);
-        setErrorMessage('');
-        callback();
-      }
-    });
-  };
-
-  const disableForm = submitState === HuntModalFormSubmitState.SUBMITTING;
-  const idPrefix = hunt ? `jr-hunt-${hunt._id}-modal-` : 'jr-hunt-new-modal-';
-  return (
-    <ModalForm
-      ref={formRef}
-      title={hunt ? 'Edit Hunt' : 'New Hunt'}
-      onSubmit={onFormSubmit}
-      submitDisabled={disableForm}
-    >
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}name`}>
-          Name
-        </FormLabel>
-        <Col xs={9}>
-          <FormControl
-            id={`${idPrefix}name`}
-            type="text"
-            value={name}
-            onChange={onNameChanged}
-            autoFocus
-            disabled={disableForm}
-          />
-        </Col>
-      </FormGroup>
-
-      <h3>Users and permissions</h3>
-
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}signup-message`}>
-          Signup message
-        </FormLabel>
-        <Col xs={9}>
-          <FormControl
-            id={`${idPrefix}signup-message`}
-            as="textarea"
-            value={signupMessage}
-            onChange={onSignupMessageChanged}
-            disabled={disableForm}
-          />
-          <FormText>
-            This message (rendered as markdown) will be shown to users who aren&apos;t part of the hunt. This is a good place to put directions for how to sign up.
-          </FormText>
-        </Col>
-      </FormGroup>
-
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}open-signups`}>
-          Open invites
-        </FormLabel>
-        <Col xs={9}>
-          <FormCheck
-            id={`${idPrefix}open-signups`}
-            checked={openSignups}
-            onChange={onOpenSignupsChanged}
-            disabled={disableForm}
-          />
-          <FormText>
-            If open invites are enabled, then any current member of the hunt can add a new member to the hunt. Otherwise, only operators can add new members.
-          </FormText>
-        </Col>
-      </FormGroup>
-
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}has-guess-queue`}>
-          Guess queue
-        </FormLabel>
-        <Col xs={9}>
-          <FormCheck
-            id={`${idPrefix}has-guess-queue`}
-            checked={hasGuessQueue}
-            onChange={onHasGuessQueueChanged}
-            disabled={disableForm}
-          />
-          <FormText>
-            If enabled, users can submit guesses for puzzles but operators must mark them as correct. If disabled, any user can enter the puzzle answer.
-          </FormText>
-        </Col>
-      </FormGroup>
-
-      <h3>Hunt website</h3>
-
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}homepage-url`}>
-          Homepage URL
-        </FormLabel>
-        <Col xs={9}>
-          <FormControl
-            id={`${idPrefix}homepage-url`}
-            type="text"
-            value={homepageUrl}
-            onChange={onHomepageUrlChanged}
-            disabled={disableForm}
-          />
-          <FormText>
-            If provided, a link to the hunt homepage will be placed on the landing page.
-          </FormText>
-        </Col>
-      </FormGroup>
-
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}submit-template`}>
-          Submit URL template
-        </FormLabel>
-        <Col xs={9}>
-          <FormControl
-            id={`${idPrefix}submit-template`}
-            type="text"
-            value={submitTemplate}
-            onChange={onSubmitTemplateChanged}
-            disabled={disableForm}
-          />
-          <FormText>
-            If provided, this
-            {' '}
-            <a href="https://mustache.github.io/mustache.5.html">Mustache template</a>
-            {' '}
-            is used to generate the link to the guess submission page. It gets as context a
-            {' '}
-            <a href="https://developer.mozilla.org/en-US/docs/Web/API/URL">parsed URL</a>
-            {', '}
-            providing variables like
-            {' '}
-            <code>hostname</code>
-            {' '}
-            or
-            {' '}
-            <code>pathname</code>
-            {'. '}
-            Because this will be used as a link directly, make sure to use &quot;triple-mustaches&quot; so that the URL components aren&apos;t escaped. As an example, setting this to
-            {' '}
-            <code>{'{{{origin}}}/submit{{{pathname}}}'}</code>
-            {' '}
-            would work for the 2018 Mystery Hunt. If not specified, the puzzle URL is used as the link to the guess submission page.
-          </FormText>
-        </Col>
-      </FormGroup>
-
-      <h3>External integrations</h3>
-
-      <FormGroup as={Row}>
-        <FormLabel column xs={3} htmlFor={`${idPrefix}mailing-lists`}>
-          Mailing lists
-        </FormLabel>
-        <Col xs={9}>
-          <FormControl
-            id={`${idPrefix}mailing-lists`}
-            type="text"
-            value={mailingLists}
-            onChange={onMailingListsChanged}
-            disabled={disableForm}
-          />
-          <FormText>
-            Users joining this hunt will be automatically added to all of these (comma-separated) lists
-          </FormText>
-        </Col>
-      </FormGroup>
-
-      {guildId ? (
-        <>
-          <FormGroup as={Row}>
-            <FormLabel column xs={3} htmlFor={`${idPrefix}puzzle-hooks-discord-channel`}>
-              Puzzle notifications Discord channel
-            </FormLabel>
-            <Col xs={9}>
-              <DiscordChannelSelector
-                id={`${idPrefix}puzzle-hooks-discord-channel`}
-                guildId={guildId}
-                disable={disableForm}
-                value={puzzleHooksDiscordChannel}
-                onChange={onPuzzleHooksDiscordChannelChanged}
-              />
-              <FormText>
-                If this field is specified, when a puzzle in this hunt is added or solved, a message will be sent to the specified channel.
-              </FormText>
-            </Col>
-          </FormGroup>
-
-          <FormGroup as={Row}>
-            <FormLabel column xs={3} htmlFor={`${idPrefix}firehose-discord-channel`}>
-              Firehose Discord channel
-            </FormLabel>
-            <Col xs={9}>
-              <DiscordChannelSelector
-                id={`${idPrefix}firehose-discord-channel`}
-                guildId={guildId}
-                disable={disableForm}
-                value={firehoseDiscordChannel}
-                onChange={onFirehoseDiscordChannelChanged}
-              />
-              <FormText>
-                If this field is specified, all chat messages written in puzzles associated with this hunt will be mirrored to the specified Discord channel.
-              </FormText>
-            </Col>
-          </FormGroup>
-
-          <FormGroup as={Row}>
-            <FormLabel column xs={3} htmlFor={`${idPrefix}member-discord-role`}>
-              Discord role for members
-            </FormLabel>
-            <Col xs={9}>
-              <DiscordRoleSelector
-                id={`${idPrefix}member-discord-role`}
-                guildId={guildId}
-                disable={disableForm}
-                value={memberDiscordRole}
-                onChange={onMemberDiscordRoleChanged}
-              />
-              <FormText>
-                If set, then members of the hunt that have linked their Discord profile are added to the specified Discord role. Note that for continuity, if this setting is changed, Jolly Roger will not touch the old role (e.g. to remove members)
-              </FormText>
-            </Col>
-          </FormGroup>
-        </>
-      ) : (
-        <Alert variant="info">
-          <FontAwesomeIcon icon={faInfo} fixedWidth />
-          Discord has not been configured, so Discord settings are disabled.
-        </Alert>
-      )}
-
-      {submitState === HuntModalFormSubmitState.FAILED && <Alert variant="danger">{errorMessage}</Alert>}
-    </ModalForm>
-  );
-});
 
 const Hunt = React.memo(({ hunt }: { hunt: HuntType }) => {
   const huntId = hunt._id;
@@ -533,13 +31,7 @@ const Hunt = React.memo(({ hunt }: { hunt: HuntType }) => {
     };
   }, [huntId]);
 
-  const editModalRef = useRef<React.ElementRef<typeof HuntModalForm>>(null);
   const deleteModalRef = useRef<ModalFormHandle>(null);
-
-  const onEdit = useCallback((state: HuntModalSubmit, callback: (error?: Error) => void) => {
-    Ansible.log('Updating hunt settings', { hunt: huntId, user: Meteor.userId(), state });
-    Meteor.call('updateHunt', huntId, state, callback);
-  }, [huntId]);
 
   const onDelete = useCallback((callback: () => void) => {
     Meteor.call('destroyHunt', huntId, (err?: Error) => {
@@ -550,12 +42,6 @@ const Hunt = React.memo(({ hunt }: { hunt: HuntType }) => {
     });
   }, [huntId]);
 
-  const showEditModal = useCallback(() => {
-    if (editModalRef.current) {
-      editModalRef.current.show();
-    }
-  }, []);
-
   const showDeleteModal = useCallback(() => {
     if (deleteModalRef.current) {
       deleteModalRef.current.show();
@@ -564,11 +50,6 @@ const Hunt = React.memo(({ hunt }: { hunt: HuntType }) => {
 
   return (
     <li>
-      <HuntModalForm
-        ref={editModalRef}
-        hunt={hunt}
-        onSubmit={onEdit}
-      />
       <ModalForm
         ref={deleteModalRef}
         title="Delete Hunt"
@@ -582,7 +63,7 @@ const Hunt = React.memo(({ hunt }: { hunt: HuntType }) => {
       </ModalForm>
       <ButtonGroup size="sm">
         {canUpdate ? (
-          <Button onClick={showEditModal} variant="outline-secondary" title="Edit hunt...">
+          <Button as={Link} to={`/hunts/${huntId}/edit`} variant="outline-secondary" title="Edit hunt...">
             <FontAwesomeIcon fixedWidth icon={faEdit} />
           </Button>
         ) : undefined}
@@ -600,7 +81,11 @@ const Hunt = React.memo(({ hunt }: { hunt: HuntType }) => {
   );
 });
 
-const CreateFixtureModal = React.forwardRef((_props: {}, forwardedRef: React.Ref<HuntModalFormHandle>) => {
+type CreateFixtureModalFormHandle = {
+  show: () => void;
+};
+
+const CreateFixtureModal = React.forwardRef((_props: {}, forwardedRef: React.Ref<CreateFixtureModalFormHandle>) => {
   const [visible, setVisible] = useState(true);
   const show = useCallback(() => setVisible(true), []);
   const hide = useCallback(() => setVisible(false), []);
@@ -657,7 +142,6 @@ const CreateFixtureModal = React.forwardRef((_props: {}, forwardedRef: React.Ref
 });
 
 const HuntListPage = () => {
-  useBreadcrumb({ title: 'Hunts', path: '/hunts' });
   const huntsLoading = useSubscribe('mongo.hunts');
   const loading = huntsLoading();
 
@@ -669,21 +153,8 @@ const HuntListPage = () => {
     };
   }, []);
 
-  const addModalRef = useRef<HuntModalFormHandle>(null);
-
-  const onAdd = useCallback((state: HuntModalSubmit, callback: (error?: Error) => void): void => {
-    Ansible.log('Creating a new hunt', { user: Meteor.userId(), state });
-    Meteor.call('createHunt', state, callback);
-  }, []);
-
-  const showAddModal = useCallback(() => {
-    if (addModalRef.current) {
-      addModalRef.current.show();
-    }
-  }, []);
-
   const [renderCreateFixtureModal, setRenderCreateFixtureModal] = useState(false);
-  const createFixtureModalRef = useRef<HuntModalFormHandle>(null);
+  const createFixtureModalRef = useRef<CreateFixtureModalFormHandle>(null);
   const showCreateFixtureModal = useCallback((e: MouseEvent) => {
     e.preventDefault();
     if (renderCreateFixtureModal && createFixtureModalRef.current) {
@@ -733,14 +204,10 @@ const HuntListPage = () => {
   return (
     <div id="jr-hunts">
       <h1>Hunts</h1>
-      <HuntModalForm
-        ref={addModalRef}
-        onSubmit={onAdd}
-      />
       {canAdd && (
         <>
-          <Button onClick={showAddModal} variant="success" size="sm" title="Add new hunt...">
-            <FontAwesomeIcon icon={faPlus} />
+          <Button as={Link} to="/hunts/new" variant="success" size="sm">
+            New hunt...
           </Button>
           {!loading && hunts.length === 0 && (
             <>
