+++ imports/client/components/Routes.tsx
@@ -9,6 +9,7 @@ import FirehosePage from './FirehosePage';
 import FirstUserForm from './FirstUserForm';
 import GuessQueuePage from './GuessQueuePage';
 import HuntApp from './HuntApp';
+import HuntListApp from './HuntListApp';
 import HuntListPage from './HuntListPage';
 import HuntProfileListPage from './HuntProfileListPage';
 import Loading from './Loading';
@@ -21,26 +22,35 @@ import RootRedirector from './RootRedirector';
 import UserInvitePage from './UserInvitePage';
 import { AuthenticatedPage, UnauthenticatedPage } from './authentication';
 
+const HuntEditPage = React.lazy(() => import('./HuntEditPage'));
 const SetupPage = React.lazy(() => import('./SetupPage'));
 const RTCDebugPage = React.lazy(() => import('./RTCDebugPage'));
 
 /* Authenticated routes - if user not logged in, get redirected to /login */
 export const AuthenticatedRouteList: RouteObject[] = [
   {
-    path: '/hunts/:huntId',
-    element: <HuntApp />,
+    path: '/hunts',
+    element: <HuntListApp />,
     children: [
-      { path: 'announcements', element: <AnnouncementsPage /> },
-      { path: 'firehose', element: <FirehosePage /> },
-      { path: 'guesses', element: <GuessQueuePage /> },
-      { path: 'hunters', element: <HuntProfileListPage /> },
-      { path: 'hunters/invite', element: <UserInvitePage /> },
-      { path: 'puzzles/:puzzleId', element: <PuzzlePage /> },
-      { path: 'puzzles', element: <PuzzleListPage /> },
-      { path: '', element: <Navigate to="puzzles" replace /> },
+      {
+        path: ':huntId',
+        element: <HuntApp />,
+        children: [
+          { path: 'announcements', element: <AnnouncementsPage /> },
+          { path: 'firehose', element: <FirehosePage /> },
+          { path: 'guesses', element: <GuessQueuePage /> },
+          { path: 'hunters', element: <HuntProfileListPage /> },
+          { path: 'hunters/invite', element: <UserInvitePage /> },
+          { path: 'puzzles/:puzzleId', element: <PuzzlePage /> },
+          { path: 'puzzles', element: <PuzzleListPage /> },
+          { path: 'edit', element: <HuntEditPage /> },
+          { path: '', element: <Navigate to="puzzles" replace /> },
+        ],
+      },
+      { path: 'new', element: <HuntEditPage /> },
+      { path: '', element: <HuntListPage /> },
     ],
   },
-  { path: '/hunts', element: <HuntListPage /> },
   { path: '/users/:userId', element: <ProfilePage /> },
   { path: '/users', element: <AllProfileListPage /> },
   { path: '/setup', element: <SetupPage /> },
